refact: move reducer para um novo arquivo e cria enum de actiontypes
